Split household agent modules
