+++ src/main/scala/com/boombustgroup/amorfati/agents/Household.scala
@@ -1,8 +1,9 @@
 package com.boombustgroup.amorfati.agents
 
+import com.boombustgroup.amorfati.agents.household.*
+import com.boombustgroup.amorfati.agents.household.HouseholdParameters.*
 import com.boombustgroup.amorfati.config.*
 import com.boombustgroup.amorfati.engine.World
-import com.boombustgroup.amorfati.engine.mechanisms.SectoralMobility
 import com.boombustgroup.amorfati.init.InitRandomness
 import com.boombustgroup.amorfati.networks.Network
 import com.boombustgroup.amorfati.random.RandomStream
@@ -56,6 +57,10 @@ object PerBankFlow:
   val zero: PerBankFlow =
     PerBankFlow(PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero)
 
+/** Public household-agent contract. Heavy monthly transition logic lives under
+  * `agents.household`; this object keeps the stable API used by the engine,
+  * tests, ledger projection, and diagnostics.
+  */
 object Household:
   type ConsumerCreditGate = (BankId, PLN, RandomStream) => Boolean
 
@@ -66,59 +71,6 @@ object Household:
 
   def countEmployed(households: Vector[State]): Int =
     households.count(isEmployed)
-
-  private inline def scaledDivRaw(numerator: BigInt, denominator: BigInt): Long =
-    if denominator == 0 then 0L
-    else
-      val scaled         = numerator * BigInt(FixedPointBase.Scale)
-      val quotient       = scaled / denominator
-      val remainder      = (scaled % denominator).abs
-      val denominatorAbs = denominator.abs
-      val twiceRemainder = remainder * 2
-      if twiceRemainder < denominatorAbs then quotient.toLong
-      else
-        val resultSign = scaled.signum * denominator.signum
-        if twiceRemainder > denominatorAbs then (quotient + resultSign).toLong
-        else if quotient % 2 == 0 then quotient.toLong
-        else (quotient + resultSign).toLong
-
-  // ---- Named constants ----
-
-  // MPC sampling bounds (Beta distribution, NBP household survey)
-  private val MpcFloor   = Share.decimal(5, 1)
-  private val MpcCeiling = Share.decimal(98, 2)
-
-  // Social network precautionary saving
-  private val NeighborDistressThreshold    = Share.decimal(30, 2) // fraction of neighbors in distress that triggers adj
-  private val NeighborDistressConsAdj      = Share.decimal(90, 2) // consumption multiplier when distress exceeds threshold
-  private val NeighborDistressRetrainBoost = Share.decimal(5, 2)  // additional retraining prob when neighbors distressed
-
-  // Labor market
-  private val UnemploymentRetrainingThreshold = 6 // months unemployed before eligible for retraining
-  private val PostFailedRetrainingMonths      = 7 // months assigned after failed retraining (duration + 1)
-
-  // Consumer credit
-  private val DisposableWageThreshold = Share.decimal(3, 1) // disposable/wage ratio below which HH may borrow
-  private val MinConsumerLoanSize     = PLN(100)            // minimum loan size (PLN)
-  private val ConsumerDebtInitFrac    = Share.decimal(3, 1) // init consumer debt as fraction of mortgage draw
-
-  // Financial-distress lifecycle behavior
-  private val LiquidityStressConsumptionMultiplier = Share.decimal(95, 2) // precautionary consumption cut after stress
-  private val ArrearsConsumptionMultiplier         = Share.decimal(90, 2) // stronger cut when obligations are in arrears
-  private val RestructuringConsumptionMultiplier   = Share.decimal(85, 2) // workout-state spending constraint
-  private val DefaultedConsumptionMultiplier       = Share.decimal(80, 2) // post-default/bankruptcy cash discipline
-
-  // Init sampling
-  private val GpwEquityInitFrac     = Share.decimal(5, 2)  // fraction of savings allocated to GPW equity at init
-  private val SectorSkillBonusCoeff = Scalar.decimal(2, 2) // coefficient for sector-specific skill bonus (log sigma)
-  private val SectorSkillBonusMax   = Scalar.decimal(1, 1) // maximum sector-specific skill bonus
-
-  // Aggregates
-  private val ImportRatioCap   = Share.decimal(65, 2) // cap on import ratio applied to goods consumption
-  private val PovertyRate50Pct = Share.decimal(50, 2) // poverty line at 50% of median income (EU AROP)
-  private val PovertyRate30Pct = Share.decimal(30, 2) // poverty line at 30% of median income (deep poverty)
-  private val ConsumptionP10   = Share.decimal(10, 2) // P10 percentile index
-  private val ConsumptionP90   = Share.decimal(90, 2) // P90 percentile index
 
   // ---- Individual household ----
 
@@ -135,7 +87,7 @@ object Household:
       bankId: BankId,                                                                     // index into the explicit bank vector (multi-bank)
       lastSectorIdx: SectorIdx,                                                           // last sector employed in (-1 = never)
       isImmigrant: Boolean,                                                               // immigrant status for wage discount + remittances
-      numDependentChildren: Int,                                                          // children ≤ 18 for 800+ social transfers
+      numDependentChildren: Int,                                                          // children <= 18 for 800+ social transfers
       education: Int,                                                                     // education level: 0=Primary, 1=Vocational, 2=Secondary, 3=Tertiary
       taskRoutineness: Share,                                                             // how routine is this worker's task bundle [0,1] (Acemoglu & Restrepo 2020)
       wageScar: Share,                                                                    // persistent wage penalty from unemployment spell (Jacobson et al. 1993)
@@ -185,18 +137,13 @@ object Household:
       LiquidityShortfallComponents(PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero)
 
   private[amorfati] def activeMortgageRemainingMonths(stocks: FinancialStocks)(using p: SimParams): Int =
-    if stocks.mortgageLoan <= PLN.Zero then 0
-    else if stocks.mortgageRemainingMonths > 0 then stocks.mortgageRemainingMonths
-    else p.housing.mortgageMaturity
+    HouseholdMortgageSchedule.activeMortgageRemainingMonths(stocks)
 
   private[amorfati] def scheduledMortgagePrincipal(stocks: FinancialStocks)(using p: SimParams): PLN =
-    val remainingMonths = activeMortgageRemainingMonths(stocks)
-    if remainingMonths <= 0 then PLN.Zero
-    else stocks.mortgageLoan / remainingMonths
+    HouseholdMortgageSchedule.scheduledMortgagePrincipal(stocks)
 
   private[amorfati] def remainingMortgageMonthsAfterPayment(stocks: FinancialStocks, closingMortgageLoan: PLN)(using p: SimParams): Int =
-    if closingMortgageLoan <= PLN.Zero then 0
-    else (activeMortgageRemainingMonths(stocks).max(1) - 1).max(1)
+    HouseholdMortgageSchedule.remainingMortgageMonthsAfterPayment(stocks, closingMortgageLoan)
 
   /** Aggregate statistics computed from individual households (Paper-06). */
   case class Aggregates(
@@ -582,11 +529,9 @@ object Household:
 
     /** Sample task routineness based on education and sector sigma.
       *
-      * High-σ sectors (BPO) have more automatable routine tasks; high education
-      * workers tend toward cognitive (non-routine) tasks. Acemoglu & Restrepo
-      * 2020.
-      *
-      * routineness = base(edu) + σ-bonus, clamped to [0.05, 0.95], with noise.
+      * High-sigma sectors (BPO) have more automatable routine tasks; high
+      * education workers tend toward cognitive (non-routine) tasks. Acemoglu &
+      * Restrepo 2020.
       */
     private[agents] def sampleTaskRoutineness(edu: Int, sectorIdx: SectorIdx, rng: RandomStream)(using p: SimParams): Share =
       val eduBase  = p.labor.sbtcEduRoutineness(edu)
@@ -594,764 +539,31 @@ object Household:
       val noise    = Scalar.randomBetween(Scalar.decimal(-25, 3), Scalar.decimal(25, 3), rng).toShare
       (eduBase + sigmaAdj + noise).clamp(Share.decimal(5, 2), Share.decimal(95, 2))
 
-  // ---- Step flow totals (immutable, folded from per-HH results) ----
-
-  /** Accumulated flow totals from one step. PLN is Long — addition is exact, no
-    * Kahan compensation needed.
-    */
-  private class StepTotals:
-    private var incomeAcc: PLN               = PLN.Zero
-    private var benefitAcc: PLN              = PLN.Zero
-    private var debtSvcAcc: PLN              = PLN.Zero
-    private var mortgagePrincipalAcc: PLN    = PLN.Zero
-    private var mortgageInterestAcc: PLN     = PLN.Zero
-    private var depIntAcc: PLN               = PLN.Zero
-    private var goodsConsAcc: PLN            = PLN.Zero
-    private var rentAcc: PLN                 = PLN.Zero
-    private var remitAcc: PLN                = PLN.Zero
-    private var pitAcc: PLN                  = PLN.Zero
-    private var socialAcc: PLN               = PLN.Zero
-    private var ccDebtSvcAcc: PLN            = PLN.Zero
-    private var ccOrigAcc: PLN               = PLN.Zero
-    private var ccApprovedOrigAcc: PLN       = PLN.Zero
-    private var ccDemandAcc: PLN             = PLN.Zero
-    private var ccRejectedAcc: PLN           = PLN.Zero
-    private var ccBankRejectedAcc: PLN       = PLN.Zero
-    private var liquidityShortfallAcc: PLN   = PLN.Zero
-    private var consumptionShortfallAcc: PLN = PLN.Zero
-    private var rentArrearsAcc: PLN          = PLN.Zero
-    private var mortgageArrearsAcc: PLN      = PLN.Zero
-    private var consumerDebtArrearsAcc: PLN  = PLN.Zero
-    private var temporaryOverdraftAcc: PLN   = PLN.Zero
-    private var ccDefaultAcc: PLN            = PLN.Zero
-    private var ccPrincipalAcc: PLN          = PLN.Zero
-    private var ccLoanDefaultAcc: PLN        = PLN.Zero
-    private var bridgeChargeOffAcc: PLN      = PLN.Zero
-    private var unmetBasicConsAcc: PLN       = PLN.Zero
-    private var discretionaryCutAcc: PLN     = PLN.Zero
-    var retrainingAttempts: Int              = 0
-    var retrainingSuccesses: Int             = 0
-    var voluntaryQuits: Int                  = 0
-
-    def add(r: HhMonthlyResult): Unit =
-      incomeAcc = incomeAcc + r.income
-      benefitAcc = benefitAcc + r.benefit
-      debtSvcAcc = debtSvcAcc + r.debtService
-      mortgagePrincipalAcc = mortgagePrincipalAcc + r.mortgagePrincipal
-      mortgageInterestAcc = mortgageInterestAcc + r.mortgageInterest
-      depIntAcc = depIntAcc + r.depositInterest
-      goodsConsAcc = goodsConsAcc + r.consumption
-      rentAcc = rentAcc + r.rent
-      remitAcc = remitAcc + r.remittance
-      pitAcc = pitAcc + r.pitTax
-      socialAcc = socialAcc + r.socialTransfer
-      ccDebtSvcAcc = ccDebtSvcAcc + r.credit.debtService
-      ccOrigAcc = ccOrigAcc + r.credit.totalOrigination
-      ccApprovedOrigAcc = ccApprovedOrigAcc + r.credit.newLoan
-      ccDemandAcc = ccDemandAcc + r.credit.creditDemand
-      ccRejectedAcc = ccRejectedAcc + r.credit.rejectedCreditDemand
-      ccBankRejectedAcc = ccBankRejectedAcc + r.credit.bankRejectedCreditDemand
-      liquidityShortfallAcc = liquidityShortfallAcc + r.credit.liquidityShortfallFinancing
-      consumptionShortfallAcc = consumptionShortfallAcc + r.credit.liquidityShortfall.consumptionShortfall
-      rentArrearsAcc = rentArrearsAcc + r.credit.liquidityShortfall.rentArrears
-      mortgageArrearsAcc = mortgageArrearsAcc + r.credit.liquidityShortfall.mortgageArrears
-      consumerDebtArrearsAcc = consumerDebtArrearsAcc + r.credit.liquidityShortfall.consumerDebtArrears
-      temporaryOverdraftAcc = temporaryOverdraftAcc + r.credit.liquidityShortfall.temporaryOverdraft
-      ccDefaultAcc = ccDefaultAcc + r.credit.defaultAmt
-      ccPrincipalAcc = ccPrincipalAcc + r.credit.principal
-      ccLoanDefaultAcc = ccLoanDefaultAcc + r.credit.consumerLoanDefault
-      bridgeChargeOffAcc = bridgeChargeOffAcc + r.credit.liquidityBridgeChargeOff
-      unmetBasicConsAcc = unmetBasicConsAcc + r.unmetBasicConsumption
-      discretionaryCutAcc = discretionaryCutAcc + r.discretionaryConsumptionCompression
-      retrainingAttempts += r.retrainingAttempt
-      retrainingSuccesses += r.retrainingSuccess
-      voluntaryQuits += r.voluntaryQuit
-
-    def income: PLN                               = incomeAcc
-    def unempBenefits: PLN                        = benefitAcc
-    def debtService: PLN                          = debtSvcAcc
-    def mortgagePrincipal: PLN                    = mortgagePrincipalAcc
-    def mortgageInterest: PLN                     = mortgageInterestAcc
-    def depositInterest: PLN                      = depIntAcc
-    def goodsConsumption: PLN                     = goodsConsAcc
-    def rent: PLN                                 = rentAcc
-    def remittances: PLN                          = remitAcc
-    def pit: PLN                                  = pitAcc
-    def socialTransfers: PLN                      = socialAcc
-    def consumerDebtService: PLN                  = ccDebtSvcAcc
-    def consumerOrigination: PLN                  = ccOrigAcc
-    def consumerApprovedOrigination: PLN          = ccApprovedOrigAcc
-    def totalConsumerCreditDemand: PLN            = ccDemandAcc
-    def totalConsumerRejectedOrigination: PLN     = ccRejectedAcc
-    def totalConsumerBankRejectedOrigination: PLN = ccBankRejectedAcc
-    def liquidityShortfallFinancing: PLN          = liquidityShortfallAcc
-    def consumptionShortfall: PLN                 = consumptionShortfallAcc
-    def rentArrears: PLN                          = rentArrearsAcc
-    def mortgageArrears: PLN                      = mortgageArrearsAcc
-    def consumerDebtArrears: PLN                  = consumerDebtArrearsAcc
-    def temporaryOverdraft: PLN                   = temporaryOverdraftAcc
-    def consumerDefault: PLN                      = ccDefaultAcc
-    def consumerPrincipal: PLN                    = ccPrincipalAcc
-    def consumerLoanDefault: PLN                  = ccLoanDefaultAcc
-    def liquidityBridgeChargeOff: PLN             = bridgeChargeOffAcc
-    def unmetBasicConsumption: PLN                = unmetBasicConsAcc
-    def discretionaryConsumptionCut: PLN          = discretionaryCutAcc
-
-  /** Build per-bank flow vector from (BankId, HhMonthlyResult) pairs. PLN is
-    * Long — addition is exact, no Kahan needed.
-    */
-  private def buildPerBankFlows(flows: Vector[(BankId, HhMonthlyResult)], nBanks: Int): Vector[PerBankFlow] =
-    val acc = Array.fill(nBanks)(PerBankFlow.zero)
-    flows.foreach: (bankId, r) =>
-      val b   = bankId.toInt
-      val cur = acc(b)
-      acc(b) = PerBankFlow(
-        income = cur.income + r.income,
-        consumption = cur.consumption + r.consumption + r.rent,
-        debtService = cur.debtService + r.debtService,
-        mortgageInterest = cur.mortgageInterest + r.mortgageInterest,
-        depositInterest = cur.depositInterest + r.depositInterest,
-        consumerDebtService = cur.consumerDebtService + r.credit.debtService,
-        consumerOrigination = cur.consumerOrigination + r.credit.totalOrigination,
-        consumerApprovedOrigination = cur.consumerApprovedOrigination + r.credit.newLoan,
-        consumerBankRejectedOrigination = cur.consumerBankRejectedOrigination + r.credit.bankRejectedCreditDemand,
-        liquidityShortfallFinancing = cur.liquidityShortfallFinancing + r.credit.liquidityShortfallFinancing,
-        consumerDefault = cur.consumerDefault + r.credit.defaultAmt,
-        consumerLoanDefault = cur.consumerLoanDefault + r.credit.consumerLoanDefault,
-        consumerPrincipal = cur.consumerPrincipal + r.credit.principal,
-      )
-    acc.toVector
-
-  // ---- Extracted per-HH pipeline types ----
-
-  /** Consumer credit result for a single household in one month. */
-  private case class CreditResult(
-      debtService: PLN,                                 // total consumer debt service (amortization + interest)
-      principal: PLN,                                   // principal component of debt service
-      interest: PLN,                                    // interest component of debt service
-      newLoan: PLN,                                     // underwritten new consumer loan, excluding same-month liquidity bridge
-      creditDemand: PLN,                                // requested underwritten consumer-credit principal before eligibility denial
-      rejectedCreditDemand: PLN,                        // requested principal rejected by borrower or bank rules
-      bankRejectedCreditDemand: PLN,                    // requested principal rejected by bank supply after household eligibility
-      creditCapacity: PLN,                              // maximum DTI-based principal capacity available this month
-      creditAccessEligible: Boolean,                    // whether stochastic access allows underwritten credit this month
-      liquidityShortfall: LiquidityShortfallComponents, // same-month bridge/write-off split by diagnostic source
-      defaultAmt: PLN,                                  // same-month bridge charge-off plus bankruptcy default amount
-      updatedDebt: PLN,                                 // outstanding consumer debt after this month's flows
-  ):
-    def liquidityShortfallFinancing: PLN = liquidityShortfall.total
-    def liquidityBridgeChargeOff: PLN    = liquidityShortfallFinancing
-    def consumerLoanDefault: PLN         = (defaultAmt - liquidityBridgeChargeOff).max(PLN.Zero)
-    // Gross flow preserves the SFC bridge leg; defaultAmt offsets bridge stock in the same month.
-    def totalOrigination: PLN            = newLoan + liquidityShortfallFinancing
-
-  /** Result of the household budget waterfall before residual settlement. */
-  private case class ConsumptionWaterfall(
-      desiredConsumption: PLN,                 // behavioral consumption before cash-priority compression
-      actualConsumption: PLN,                  // consumption paid before residual bridge/default settlement
-      unmetBasicConsumption: PLN,              // non-discretionary consumption need not covered by cash
-      discretionaryConsumptionCompression: PLN, // discretionary consumption cut before bridge/default
-  )
-
-  /** Per-HH monthly result — updated state + all flow variables for
-    * aggregation.
-    */
-  private case class HhMonthlyResult(
-      newState: State,                         // updated household state
-      income: PLN,                             // gross income net of PIT plus social transfers
-      benefit: PLN,                            // unemployment benefit component
-      consumption: PLN,                        // total consumption (goods + wealth effect, before rent)
-      debtService: PLN,                        // secured (mortgage) debt service
-      mortgagePrincipal: PLN,                  // mortgage principal component of secured debt service
-      mortgageInterest: PLN,                   // mortgage interest component of secured debt service
-      depositInterest: PLN,                    // deposit interest received
-      remittance: PLN,                         // remittance sent abroad (immigrants only)
-      pitTax: PLN,                             // PIT paid
-      socialTransfer: PLN,                     // 800+ social transfer received
-      credit: CreditResult,                    // consumer credit flows
-      voluntaryQuit: Int,                      // 1 if voluntary cross-sector quit, 0 otherwise
-      retrainingAttempt: Int,                  // 1 if retraining attempted, 0 otherwise
-      retrainingSuccess: Int,                  // 1 if retraining succeeded, 0 otherwise
-      equityWealth: PLN,                       // updated equity wealth after revaluation
-      rent: PLN,                               // monthly rent payment
-      financialStocks: FinancialStocks,
-      unmetBasicConsumption: PLN,              // basic consumption need not covered by cash
-      discretionaryConsumptionCompression: PLN, // discretionary consumption cut before bridge/default
-  )
-
-  // ---- Logic ----
-
-  /** Monthly PIT: progressive Polish brackets (12%/32%), minus kwota wolna. PIT
-    * base = gross income − ZUS employee contribution (Art. 26 ustawy o PIT).
-    */
+  /** Monthly PIT using the model's progressive Polish tax brackets. */
   def computeMonthlyPit(monthlyIncome: PLN)(using p: SimParams): PLN =
-    if monthlyIncome <= PLN.Zero then PLN.Zero
-    else
-      val afterZus   = monthlyIncome - monthlyIncome * p.social.zusEmployeeRate
-      val annualized = afterZus * Multiplier(12)
-      val grossTax   =
-        if annualized <= p.fiscal.pitBracket1Annual then annualized * p.fiscal.pitRate1
-        else
-          p.fiscal.pitBracket1Annual * p.fiscal.pitRate1 +
-            (annualized - p.fiscal.pitBracket1Annual) * p.fiscal.pitRate2
-      (grossTax - p.fiscal.pitTaxCreditAnnual).max(PLN.Zero) / 12L
+    HouseholdIncomeConstruction.computeMonthlyPit(monthlyIncome)
 
-  /** Compute 800+ social transfer (PIT-exempt, lump-sum per child ≤ 18). */
+  /** Child-dependent household social transfer. */
   def computeSocialTransfer(numChildren: Int)(using p: SimParams): PLN =
-    if numChildren <= 0 then PLN.Zero
-    else numChildren * p.fiscal.social800
+    HouseholdIncomeConstruction.computeSocialTransfer(numChildren)
 
-  /** Unemployment benefit (zasilek): 1500 PLN m1-3, 1200 PLN m4-6, 0 after. */
+  /** Unemployment benefit from spell duration. */
   def computeBenefit(monthsUnemployed: Int)(using p: SimParams): PLN =
-    if monthsUnemployed <= p.fiscal.govBenefitDuration / 2 then p.fiscal.govBenefitM1to3
-    else if monthsUnemployed <= p.fiscal.govBenefitDuration then p.fiscal.govBenefitM4to6
-    else PLN.Zero
+    HouseholdIncomeConstruction.computeBenefit(monthsUnemployed)
 
-  /** Voluntary cross-sector search for employed HH → (newStatus, quitFlag). */
-  private def tryVoluntarySearch(
-      financialStocks: FinancialStocks,
-      status: HhStatus.Employed,
-      sectorWages: Vector[PLN],
-      sectorVacancies: Vector[Int],
-      rng: RandomStream,
-  )(using p: SimParams): (HhStatus, Int) =
-    if !p.labor.voluntarySearchProb.sampleBelow(rng) then return (status, 0)
-    val targetSector      =
-      SectoralMobility.selectTargetSector(status.sectorIdx.toInt, sectorWages, sectorVacancies, p.labor.frictionMatrix, p.labor.vacancyWeight, rng)
-    if sectorVacancies(targetSector) <= 0 then return (status, 0)
-    val targetAvgWage     = sectorWages(targetSector)
-    val friction          = p.labor.frictionMatrix(status.sectorIdx.toInt)(targetSector)
-    val frictionNetWage   = targetAvgWage * SectoralMobility.crossSectorWagePenalty(friction)
-    val wageThresholdMult = Multiplier.One + p.labor.voluntaryWageThreshold.toMultiplier
-    if frictionNetWage <= status.wage * wageThresholdMult then return (status, 0)
-    if friction <= p.labor.adjacentFrictionMax then
-      val rp = SectoralMobility.frictionAdjustedParams(friction, p.labor.frictionDurationMult, p.labor.frictionCostMult)
-      if financialStocks.demandDeposit > rp.cost then (HhStatus.Retraining(rp.duration, SectorIdx(targetSector), rp.cost), 1)
-      else (status, 0)
-    else (status, 0)
+  /** Public test hook for the buffer-stock MPC rule. */
+  private[amorfati] def updateMpc(hh: State, financialStocks: FinancialStocks, income: PLN, status: HhStatus)(using p: SimParams): Share =
+    HouseholdDistressMachine.updateMpc(hh, financialStocks, income, status)
 
-  /** Retraining for unemployed HH → (newStatus, attemptFlag, successFlag). */
-  private def tryRetraining(
-      hh: State,
+  /** Public test hook for controlled savings-buffer drawdown. */
+  private[amorfati] def computeSavingsDrawdown(
       financialStocks: FinancialStocks,
+      cashOnHand: PLN,
       status: HhStatus,
-      neighborDistress: Share,
-      sectorWages: Option[Vector[PLN]],
-      sectorVacancies: Option[Vector[Int]],
-      rng: RandomStream,
-  )(using p: SimParams): (HhStatus, Int, Int) =
-    status match
-      case HhStatus.Unemployed(months) if months > UnemploymentRetrainingThreshold && p.household.retrainingEnabled =>
-        val retrainProb = p.household.retrainingProb +
-          (if neighborDistress > NeighborDistressThreshold then NeighborDistressRetrainBoost else Share.Zero)
-        if financialStocks.demandDeposit > p.household.retrainingCost && retrainProb.sampleBelow(rng) then
-          if sectorWages.isDefined then
-            val sw           = sectorWages.get
-            val sv           = sectorVacancies.get
-            val fromSector   = if hh.lastSectorIdx.toInt >= 0 then hh.lastSectorIdx.toInt else 0
-            val targetSector = SectoralMobility.selectTargetSector(fromSector, sw, sv, p.labor.frictionMatrix, p.labor.vacancyWeight, rng)
-            val friction     = p.labor.frictionMatrix(fromSector)(targetSector)
-            val rp           = SectoralMobility.frictionAdjustedParams(friction, p.labor.frictionDurationMult, p.labor.frictionCostMult)
-            if financialStocks.demandDeposit > rp.cost then (HhStatus.Retraining(rp.duration, SectorIdx(targetSector), rp.cost), 1, 0)
-            else (status, 0, 0)
-          else
-            val targetSector = rng.nextInt(p.sectorDefs.length)
-            (HhStatus.Retraining(p.household.retrainingDuration, SectorIdx(targetSector), p.household.retrainingCost), 1, 0)
-        else (status, 0, 0)
+  )(using p: SimParams): PLN =
+    HouseholdDistressMachine.computeSavingsDrawdown(financialStocks, cashOnHand, status)
 
-      case HhStatus.Retraining(monthsLeft, targetSector, cost) =>
-        if monthsLeft <= 1 then
-          val afterSkill      = applySkillDecay(hh, status)
-          val afterHealth     = applyHealthScarring(hh, status)
-          // baseSuccess × skill × (1 − healthPenalty) × eduMultiplier
-          val skillHealthProb = p.household.retrainingBaseSuccess * afterSkill * (Share.One - afterHealth)
-          val eduMult         = p.social.eduRetrainMultiplier(hh.education)
-          val baseSuccessProb = (skillHealthProb * eduMult).toShare // Multiplier → Share (probability)
-          val successProb     =
-            val fromSector = if hh.lastSectorIdx.toInt >= 0 then hh.lastSectorIdx.toInt else 0
-            val friction   = p.labor.frictionMatrix(fromSector)(targetSector.toInt)
-            baseSuccessProb * (Share.One - friction * SectoralMobility.FrictionSuccessDiscount)
-          if successProb.sampleBelow(rng) then (HhStatus.Unemployed(0), 0, 1)
-          else (HhStatus.Unemployed(PostFailedRetrainingMonths), 0, 0)
-        else (HhStatus.Retraining(monthsLeft - 1, targetSector, cost), 0, 0)
-
-      case _ => (status, 0, 0)
-
-  private def consumerLoanCapacity(monthlyPaymentHeadroom: PLN, paymentFactor: Rate): PLN =
-    if monthlyPaymentHeadroom <= PLN.Zero || paymentFactor <= Rate.Zero then PLN.Zero
-    else monthlyPaymentHeadroom / paymentFactor.toMultiplier
-
-  // First-month liquidity stress still goes through normal underwriting;
-  // repeated arrears/default states close access to new unsecured credit.
-  private def consumerCreditBlockedByDistress(state: HhFinancialDistressState): Boolean =
-    state match
-      case HhFinancialDistressState.Arrears | HhFinancialDistressState.Restructuring | HhFinancialDistressState.Defaulted |
-          HhFinancialDistressState.Bankruptcy =>
-        true
-      case HhFinancialDistressState.Current | HhFinancialDistressState.LiquidityStress =>
-        false
-
-  private def financialDistressConsumptionMultiplier(state: HhFinancialDistressState): Share =
-    state match
-      case HhFinancialDistressState.Current         => Share.One
-      case HhFinancialDistressState.LiquidityStress => LiquidityStressConsumptionMultiplier
-      case HhFinancialDistressState.Arrears         => ArrearsConsumptionMultiplier
-      case HhFinancialDistressState.Restructuring   => RestructuringConsumptionMultiplier
-      case HhFinancialDistressState.Defaulted       => DefaultedConsumptionMultiplier
-      case HhFinancialDistressState.Bankruptcy      => DefaultedConsumptionMultiplier
-
-  private def applyFinancialDistressConsumptionAdjustment(consumption: PLN, state: HhFinancialDistressState)(using p: SimParams): PLN =
-    val multiplier = financialDistressConsumptionMultiplier(state)
-    if multiplier >= Share.One || consumption <= PLN.Zero then consumption
-    else (consumption * multiplier).max(p.household.basicConsumptionFloor.min(consumption))
-
-  // Recovery is staged so a household that exits arrears/default spends one
-  // month in constrained restructuring before returning to Current.
-  private def recoveryDistressState(previous: HhFinancialDistressState): HhFinancialDistressState =
-    previous match
-      case HhFinancialDistressState.Arrears | HhFinancialDistressState.Defaulted | HhFinancialDistressState.Bankruptcy          =>
-        HhFinancialDistressState.Restructuring
-      case HhFinancialDistressState.Current | HhFinancialDistressState.LiquidityStress | HhFinancialDistressState.Restructuring =>
-        HhFinancialDistressState.Current
-
-  // The default threshold and legal write-off threshold are separate: arrears
-  // can enter Defaulted before consumer debt is fully written off.
-  private def advanceFinancialDistressState(
-      previous: HhFinancialDistressState,
-      status: HhStatus,
-      distressMonths: Int,
-      personalInsolvency: Boolean,
-  )(using p: SimParams): HhFinancialDistressState =
-    if status == HhStatus.Bankrupt || personalInsolvency then HhFinancialDistressState.Bankruptcy
-    else if distressMonths <= 0 then recoveryDistressState(previous)
-    else if distressMonths >= p.household.bankruptcyDistressMonths then HhFinancialDistressState.Defaulted
-    else if distressMonths >= 2 then HhFinancialDistressState.Arrears
-    else HhFinancialDistressState.LiquidityStress
-
-  private def applyConsumerCreditGate(
-      hh: State,
-      demand: PLN,
-      householdEligible: Boolean,
-      rng: RandomStream,
-      consumerCreditGate: Option[ConsumerCreditGate],
-  ): (PLN, PLN, PLN) =
-    if demand <= PLN.Zero then (PLN.Zero, PLN.Zero, PLN.Zero)
-    else if !householdEligible then (PLN.Zero, demand, PLN.Zero)
-    else if consumerCreditGate.forall(gate => gate(hh.bankId, demand, rng)) then (demand, PLN.Zero, PLN.Zero)
-    else (PLN.Zero, demand, demand)
-
-  /** Consumer credit for one HH: debt service, origination, principal. */
-  private def processConsumerCredit(
-      hh: State,
-      financialStocks: FinancialStocks,
-      income: PLN,
-      disposable: PLN,
-      debtService: PLN,
-      world: World,
-      bankRates: Option[BankRates],
-      rng: RandomStream,
-      consumerCreditGate: Option[ConsumerCreditGate],
-  )(using p: SimParams): CreditResult =
-    val consumerRate: Rate = bankRates match
-      case Some(br) => br.lendingRates(hh.bankId.toInt) + p.household.ccSpread
-      case None     => world.nbp.referenceRate + p.household.ccSpread
-    val consumerPrin       = financialStocks.consumerLoan * p.household.ccAmortRate
-    val consumerInterest   = financialStocks.consumerLoan * consumerRate.monthly
-    val consumerDebtSvc    = consumerPrin + consumerInterest
-    val paymentFactor      = p.household.ccAmortRate + consumerRate.monthly
-
-    val (creditDemand, rejectedCreditDemand, bankRejectedCreditDemand, newConsumerLoan, creditCapacity, creditAccessEligible) = hh.status match
-      case HhStatus.Employed(_, _, wage)                                             =>
-        val stressed = disposable < wage * DisposableWageThreshold
-        val eligible = stressed && !consumerCreditBlockedByDistress(hh.financialDistressState) && p.household.ccEligRate.sampleBelow(rng)
-        if !stressed then (PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, false)
-        else
-          val totalDbtSvc                        = debtService + consumerDebtSvc
-          val existingDti                        = if income > PLN.Zero then (totalDbtSvc / income).toShare else Share.One
-          val headroomShare                      = (p.household.ccMaxDti - existingDti).max(Share.Zero)
-          val paymentHeadroom                    = income * headroomShare
-          val principalCapacity                  = consumerLoanCapacity(paymentHeadroom, paymentFactor).min(p.household.ccMaxLoan)
-          val stressGap                          = (wage * DisposableWageThreshold - disposable).max(PLN.Zero)
-          val essentialGap                       = (p.household.basicConsumptionFloor + consumerDebtSvc - disposable).max(PLN.Zero)
-          val desired                            = stressGap.max(essentialGap).min(principalCapacity)
-          val demand                             = if desired > MinConsumerLoanSize then desired else PLN.Zero
-          val (approved, rejected, bankRejected) = applyConsumerCreditGate(hh, demand, eligible, rng, consumerCreditGate)
-          (demand, rejected, bankRejected, approved, principalCapacity, eligible)
-      case HhStatus.Unemployed(_) | HhStatus.Retraining(_, _, _) | HhStatus.Bankrupt =>
-        (PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, PLN.Zero, false)
-
-    val updatedDebt = (financialStocks.consumerLoan + newConsumerLoan - consumerPrin).max(PLN.Zero)
-
-    CreditResult(
-      debtService = consumerDebtSvc,
-      principal = consumerPrin,
-      interest = consumerInterest,
-      newLoan = newConsumerLoan,
-      creditDemand = creditDemand,
-      rejectedCreditDemand = rejectedCreditDemand,
-      bankRejectedCreditDemand = bankRejectedCreditDemand,
-      creditCapacity = creditCapacity,
-      creditAccessEligible = creditAccessEligible,
-      liquidityShortfall = LiquidityShortfallComponents.Zero,
-      defaultAmt = PLN.Zero,
-      updatedDebt = updatedDebt,
-    )
-
-  /** Intermediate result after income/consumption pipeline, before branching.
-    */
-  private case class MonthlyFlows(
-      hh: State,
-      financialStocks: FinancialStocks,
-      income: PLN,
-      benefit: PLN,
-      newStatus: HhStatus,
-      debtService: PLN,
-      mortgagePrincipal: PLN,
-      mortgageInterest: PLN,
-      depositInterest: PLN,
-      remittance: PLN,
-      pitTax: PLN,
-      socialTransfer: PLN,
-      credit: CreditResult,
-      desiredConsumption: PLN,
-      consumption: PLN,
-      unmetBasicConsumption: PLN,
-      discretionaryConsumptionCompression: PLN,
-      newEquityWealth: PLN,
-      newSavings: PLN, // raw closing liquid balance before non-negative deposit settlement
-      newDebt: PLN,
-      newMortgageRemainingMonths: Int,
-      neighborDistress: Share,
-  )
-
-  /** Pay non-discretionary consumption first, then obligations, then only the
-    * affordable discretionary part. Unpaid basic need is deprivation, not
-    * bridge debt; unpaid obligations still flow to the residual settlement
-    * diagnostics.
-    */
-  private def applyConsumptionWaterfall(
-      openingDemandDeposit: PLN,
-      income: PLN,
-      approvedConsumerLoan: PLN,
-      obligations: PLN,
-      desiredConsumption: PLN,
-  )(using p: SimParams): ConsumptionWaterfall =
-    if desiredConsumption <= PLN.Zero then ConsumptionWaterfall(desiredConsumption, desiredConsumption, PLN.Zero, PLN.Zero)
-    else
-      val desiredPositive      = desiredConsumption.max(PLN.Zero)
-      val basicNeed            = p.household.basicConsumptionFloor.min(desiredPositive)
-      val availableCash        = (openingDemandDeposit + income + approvedConsumerLoan).max(PLN.Zero)
-      val paidBasic            = basicNeed.min(availableCash)
-      val availableAfterBasic  = (availableCash - paidBasic).max(PLN.Zero)
-      val availableAfterBills  = (availableAfterBasic - obligations.max(PLN.Zero)).max(PLN.Zero)
-      val desiredDiscretionary = (desiredPositive - basicNeed).max(PLN.Zero)
-      val paidDiscretionary    = desiredDiscretionary.min(availableAfterBills)
-      ConsumptionWaterfall(
-        desiredConsumption = desiredConsumption,
-        actualConsumption = paidBasic + paidDiscretionary,
-        unmetBasicConsumption = basicNeed - paidBasic,
-        discretionaryConsumptionCompression = desiredDiscretionary - paidDiscretionary,
-      )
-
-  private def financialDistressTriggered(f: MonthlyFlows)(using p: SimParams): Boolean =
-    val essentialOutflows = (f.hh.monthlyRent + f.debtService + f.credit.debtService).max(f.hh.monthlyRent)
-    val configuredFloor   = essentialOutflows * p.household.bankruptcyThreshold.toMultiplier
-    f.newSavings < PLN.Zero || f.unmetBasicConsumption > PLN.Zero || f.newSavings < configuredFloor
-
-  private def underwriteResidualShortfall(f: MonthlyFlows, rng: RandomStream, consumerCreditGate: Option[ConsumerCreditGate]): MonthlyFlows =
-    val shortfall         = (PLN.Zero - f.newSavings).max(PLN.Zero)
-    val remainingCapacity = (f.credit.creditCapacity - f.credit.newLoan).max(PLN.Zero)
-    val desired           = shortfall.min(remainingCapacity)
-    val demand            = if desired > MinConsumerLoanSize then desired else PLN.Zero
-    if demand <= PLN.Zero then f
-    else
-      val (approved, rejected, bankRejected) = applyConsumerCreditGate(f.hh, demand, f.credit.creditAccessEligible, rng, consumerCreditGate)
-      f.copy(
-        credit = f.credit.copy(
-          newLoan = f.credit.newLoan + approved,
-          creditDemand = f.credit.creditDemand + demand,
-          rejectedCreditDemand = f.credit.rejectedCreditDemand + rejected,
-          bankRejectedCreditDemand = f.credit.bankRejectedCreditDemand + bankRejected,
-          updatedDebt = f.credit.updatedDebt + approved,
-        ),
-        newSavings = f.newSavings + approved,
-      )
-
-  /** Per-HH monthly pipeline: income → tax → credit → consumption → equity. */
-  private def computeMonthlyFlows(
-      hh: State,
-      financialStocks: FinancialStocks,
-      world: World,
-      rng: RandomStream,
-      bankRates: Option[BankRates],
-      equityIndexReturn: Rate,
-      distressedIds: java.util.BitSet,
-      consumerCreditGate: Option[ConsumerCreditGate],
-  )(using p: SimParams): MonthlyFlows =
-    val (baseIncome, benefit, newStatus) = computeIncome(hh)
-
-    // Mortgage service uses the contract's remaining maturity for principal and bank rates for interest.
-    val mortgageRate: Rate = bankRates match
-      case Some(br) => br.lendingRates(hh.bankId.toInt)
-      case None     => world.nbp.referenceRate + p.housing.mortgageSpread
-
-    // Deposit interest (monetary transmission channel 2)
-    val depInterest: PLN = bankRates match
-      case Some(br) => financialStocks.demandDeposit * br.depositRates(hh.bankId.toInt).monthly
-      case None     => PLN.Zero
-
-    val grossIncome       = baseIncome + depInterest.max(PLN.Zero)
-    val pitTax            = computeMonthlyPit(grossIncome)
-    val socialTransfer    = computeSocialTransfer(hh.numDependentChildren)
-    val income            = grossIncome - pitTax + socialTransfer
-    val mortgagePrincipal = scheduledMortgagePrincipal(financialStocks)
-    val mortgageInterest  = financialStocks.mortgageLoan * mortgageRate.max(Rate.Zero).monthly
-    val thisDebtService   = mortgagePrincipal + mortgageInterest
-    val newMortgageLoan   = (financialStocks.mortgageLoan - mortgagePrincipal).max(PLN.Zero)
-
-    val remittance =
-      if hh.isImmigrant then income * p.immigration.remitRate
-      else PLN.Zero
-
-    val obligations         = hh.monthlyRent + thisDebtService + remittance
-    val disposablePreCredit = (income - obligations).max(PLN.Zero)
-    val credit              = processConsumerCredit(hh, financialStocks, income, disposablePreCredit, thisDebtService, world, bankRates, rng, consumerCreditGate)
-    val fullObligations     = obligations + credit.debtService
-    val disposable          = (income - fullObligations).max(PLN.Zero)
-    val savingsDrawdown     = computeSavingsDrawdown(financialStocks, income, newStatus)
-    val consumptionBudget   = disposable + credit.newLoan + savingsDrawdown
-    val consumption         = consumptionBudget * hh.mpc
-
-    // Social network precautionary effect
-    val neighborDistress            = neighborDistressRatioFast(hh, distressedIds)
-    val consumptionAdj              =
-      if neighborDistress > NeighborDistressThreshold then consumption * NeighborDistressConsAdj else consumption
-    val distressAdjustedConsumption = applyFinancialDistressConsumptionAdjustment(consumptionAdj, hh.financialDistressState)
-
-    // Wealth effects: equity (GPW) + housing (Meen HPI)
-    val newEquityWealth       = (financialStocks.equity * (Multiplier.One + equityIndexReturn.toMultiplier)).max(PLN.Zero)
-    val equityGain            = newEquityWealth - financialStocks.equity
-    val equityBoost           =
-      if equityGain > PLN.Zero then equityGain * p.equity.wealthEffectMpc
-      else PLN.Zero
-    val housingBoost          = world.real.housing.lastWealthEffect / world.derivedTotalPopulation.toLong.max(1L)
-    val consumptionWithWealth = distressAdjustedConsumption + equityBoost + housingBoost
-    val consumptionWaterfall  =
-      applyConsumptionWaterfall(financialStocks.demandDeposit, income, credit.newLoan, fullObligations, consumptionWithWealth)
-
-    val initialFlows = MonthlyFlows(
-      hh = hh,
-      financialStocks = financialStocks,
-      income = income,
-      benefit = benefit,
-      newStatus = newStatus,
-      debtService = thisDebtService,
-      mortgagePrincipal = mortgagePrincipal,
-      mortgageInterest = mortgageInterest,
-      depositInterest = depInterest.max(PLN.Zero),
-      remittance = remittance,
-      pitTax = pitTax,
-      socialTransfer = socialTransfer,
-      credit = credit,
-      desiredConsumption = consumptionWaterfall.desiredConsumption,
-      consumption = consumptionWaterfall.actualConsumption,
-      unmetBasicConsumption = consumptionWaterfall.unmetBasicConsumption,
-      discretionaryConsumptionCompression = consumptionWaterfall.discretionaryConsumptionCompression,
-      newEquityWealth = newEquityWealth,
-      newSavings = financialStocks.demandDeposit + income - fullObligations + credit.newLoan - consumptionWaterfall.actualConsumption,
-      newDebt = newMortgageLoan,
-      newMortgageRemainingMonths = remainingMortgageMonthsAfterPayment(financialStocks, newMortgageLoan),
-      neighborDistress = neighborDistress,
-    )
-    underwriteResidualShortfall(initialFlows, rng, consumerCreditGate)
-
-  /** Resolve flows into final HhMonthlyResult: bankruptcy or survival branch.
-    */
-  private def processHousehold(
-      hh: State,
-      financialStocks: FinancialStocks,
-      world: World,
-      rng: RandomStream,
-      bankRates: Option[BankRates],
-      equityIndexReturn: Rate,
-      sectorWages: Option[Vector[PLN]],
-      sectorVacancies: Option[Vector[Int]],
-      distressedIds: java.util.BitSet,
-      consumerCreditGate: Option[ConsumerCreditGate],
-  )(using p: SimParams): HhMonthlyResult =
-    val f              = computeMonthlyFlows(hh, financialStocks, world, rng, bankRates, equityIndexReturn, distressedIds, consumerCreditGate)
-    val distressMonths =
-      if financialDistressTriggered(f) then hh.financialDistressMonths + 1 else 0
-    if distressMonths > p.household.personalInsolvencyDistressMonths then resolveBankruptcy(f, distressMonths)
-    else resolveSurvival(f, sectorWages, sectorVacancies, rng, distressMonths)
-
-  /** Diagnostic-only attribution: outflows consume available cash in a stable
-    * audit priority, and any unfunded tail becomes the component split of the
-    * same-month liquidity bridge/default diagnostic.
-    */
-  private def attributeLiquidityShortfall(f: MonthlyFlows, rawDemandDeposit: PLN, temporaryOutflow: PLN): LiquidityShortfallComponents =
-    if rawDemandDeposit >= PLN.Zero then LiquidityShortfallComponents.Zero
-    else
-      var available = f.financialStocks.demandDeposit + f.income + f.credit.newLoan
-
-      def gap(outflow: PLN): PLN =
-        if outflow <= PLN.Zero then
-          available = available - outflow
-          PLN.Zero
-        else if available >= outflow then
-          available = available - outflow
-          PLN.Zero
-        else
-          val shortfall = outflow - available.max(PLN.Zero)
-          available = PLN.Zero
-          shortfall
-
-      val components = LiquidityShortfallComponents(
-        consumptionShortfall = gap(f.consumption),
-        rentArrears = gap(f.hh.monthlyRent),
-        mortgageArrears = gap(f.debtService),
-        consumerDebtArrears = gap(f.credit.debtService),
-        temporaryOverdraft = gap(f.remittance + temporaryOutflow),
-      )
-      val expected   = PLN.Zero - rawDemandDeposit
-      require(
-        components.total == expected,
-        s"Household liquidity shortfall components must sum to settlement shortfall, got ${components.total} and $expected",
-      )
-      components
-
-  private def settleLiquidityShortfall(rawDemandDeposit: PLN, credit: CreditResult, components: LiquidityShortfallComponents): (PLN, CreditResult) =
-    if rawDemandDeposit >= PLN.Zero then (rawDemandDeposit, credit)
-    else
-      val shortfall = PLN.Zero - rawDemandDeposit
-      require(
-        components.total == shortfall,
-        s"Household liquidity shortfall settlement requires components to match total, got ${components.total} and $shortfall",
-      )
-      (
-        PLN.Zero,
-        // Residual gaps are explicit same-month bridge defaults, not new ordinary consumer-loan stock.
-        credit.copy(
-          liquidityShortfall = credit.liquidityShortfall + components,
-          defaultAmt = credit.defaultAmt + shortfall,
-        ),
-      )
-
-  /** Personal-insolvency branch: write off consumer debt and equity without
-    * removing the household from the labor force.
-    */
-  private def resolveBankruptcy(f: MonthlyFlows, distressMonths: Int)(using p: SimParams): HhMonthlyResult =
-    // Consumer credit stock is reduced earlier by same-month principal only.
-    // Default the remaining balance so bankruptcy stays aligned with the stock
-    // identity used by BankingEconomics/SFC.
-    val liquidityShortfall                  = attributeLiquidityShortfall(f, f.newSavings, temporaryOutflow = PLN.Zero)
-    val (finalDemandDeposit, settledCredit) = settleLiquidityShortfall(f.newSavings, f.credit, liquidityShortfall)
-    val ccDefaultAmt                        = settledCredit.defaultAmt + settledCredit.updatedDebt
-    val creditWithDef                       = settledCredit.copy(defaultAmt = ccDefaultAmt, updatedDebt = PLN.Zero)
-    val financial                           = FinancialStocks(
-      demandDeposit = finalDemandDeposit,
-      mortgageLoan = f.newDebt,
-      consumerLoan = PLN.Zero,
-      equity = PLN.Zero,
-      mortgageRemainingMonths = f.newMortgageRemainingMonths,
-    )
-    HhMonthlyResult(
-      newState = f.hh.copy(
-        status = f.newStatus,
-        financialDistressMonths = 0,
-        financialDistressState = advanceFinancialDistressState(f.hh.financialDistressState, f.newStatus, distressMonths, personalInsolvency = true),
-      ),
-      income = f.income,
-      benefit = f.benefit,
-      consumption = f.consumption,
-      debtService = f.debtService,
-      mortgagePrincipal = f.mortgagePrincipal,
-      mortgageInterest = f.mortgageInterest,
-      depositInterest = f.depositInterest,
-      remittance = f.remittance,
-      pitTax = f.pitTax,
-      socialTransfer = f.socialTransfer,
-      credit = creditWithDef,
-      voluntaryQuit = 0,
-      retrainingAttempt = 0,
-      retrainingSuccess = 0,
-      equityWealth = PLN.Zero,
-      rent = f.hh.monthlyRent,
-      financialStocks = financial,
-      unmetBasicConsumption = f.unmetBasicConsumption,
-      discretionaryConsumptionCompression = f.discretionaryConsumptionCompression,
-    )
-
-  /** Survival branch: skill decay, labor transitions, state update. */
-  private def resolveSurvival(
-      f: MonthlyFlows,
-      sectorWages: Option[Vector[PLN]],
-      sectorVacancies: Option[Vector[Int]],
-      rng: RandomStream,
-      distressMonths: Int,
-  )(using p: SimParams): HhMonthlyResult =
-    val afterSkill    = applySkillDecay(f.hh, f.newStatus)
-    val afterHealth   = applyHealthScarring(f.hh, f.newStatus)
-    val afterWageScar = applyWageScar(f.hh, f.newStatus)
-    val afterMpc      = updateMpc(f.hh, f.financialStocks, f.income, f.newStatus)
-
-    val (afterVoluntary, vQuit) = f.newStatus match
-      case emp: HhStatus.Employed if sectorWages.isDefined =>
-        tryVoluntarySearch(f.financialStocks, emp, sectorWages.get, sectorVacancies.get, rng)
-      case _                                               => (f.newStatus, 0)
-
-    val (finalStatus, rAttempt, rSuccess) =
-      tryRetraining(f.hh, f.financialStocks, afterVoluntary, f.neighborDistress, sectorWages, sectorVacancies, rng)
-    val nextFinancialDistressState        =
-      advanceFinancialDistressState(f.hh.financialDistressState, finalStatus, distressMonths, personalInsolvency = false)
-
-    val retrainingCostThisMonth             = finalStatus match
-      case HhStatus.Retraining(ml, _, cost) if ml == p.household.retrainingDuration - 1 => cost
-      case _                                                                            => PLN.Zero
-    val rawFinalDemandDeposit               = f.newSavings - retrainingCostThisMonth
-    val liquidityShortfall                  = attributeLiquidityShortfall(f, rawFinalDemandDeposit, temporaryOutflow = retrainingCostThisMonth)
-    val (finalDemandDeposit, settledCredit) =
-      settleLiquidityShortfall(rawFinalDemandDeposit, f.credit, liquidityShortfall)
-    val financial                           = FinancialStocks(
-      demandDeposit = finalDemandDeposit,
-      mortgageLoan = f.newDebt,
-      consumerLoan = settledCredit.updatedDebt,
-      equity = f.newEquityWealth,
-      mortgageRemainingMonths = f.newMortgageRemainingMonths,
-    )
-
-    HhMonthlyResult(
-      newState = f.hh.copy(
-        skill = afterSkill,
-        healthPenalty = afterHealth,
-        wageScar = afterWageScar,
-        mpc = afterMpc,
-        status = finalStatus,
-        financialDistressMonths = distressMonths,
-        financialDistressState = nextFinancialDistressState,
-      ),
-      income = f.income,
-      benefit = f.benefit,
-      consumption = f.consumption,
-      debtService = f.debtService,
-      mortgagePrincipal = f.mortgagePrincipal,
-      mortgageInterest = f.mortgageInterest,
-      depositInterest = f.depositInterest,
-      remittance = f.remittance,
-      pitTax = f.pitTax,
-      socialTransfer = f.socialTransfer,
-      credit = settledCredit,
-      voluntaryQuit = vQuit,
-      retrainingAttempt = rAttempt,
-      retrainingSuccess = rSuccess,
-      equityWealth = f.newEquityWealth,
-      rent = f.hh.monthlyRent,
-      financialStocks = financial,
-      unmetBasicConsumption = f.unmetBasicConsumption,
-      discretionaryConsumptionCompression = f.discretionaryConsumptionCompression,
-    )
-
-  /** Monthly entry point: map processHousehold + accumulate + aggregate. */
+  /** Monthly entry point for the household agent. */
   def step(
       households: Vector[State],
       financialStocks: Vector[FinancialStocks],
@@ -1367,195 +579,23 @@ object Household:
       sectorVacancies: Option[Vector[Int]] = None,
       consumerCreditGate: Option[ConsumerCreditGate] = None,
   )(using p: SimParams): StepResult =
-    require(
-      households.length == financialStocks.length,
-      s"Household.step requires aligned households and financialStocks, got ${households.length} households and ${financialStocks.length} stock rows",
+    HouseholdStepRunner.step(
+      households,
+      financialStocks,
+      world,
+      marketWage,
+      reservationWage,
+      importAdj,
+      rng,
+      nBanks,
+      bankRates,
+      equityIndexReturn,
+      sectorWages,
+      sectorVacancies,
+      consumerCreditGate,
     )
-    require(
-      financialStocks.forall(_.demandDeposit >= PLN.Zero),
-      "Household.step requires non-negative opening demandDeposit balances; liquidity shortfalls must enter as consumer-loan stocks",
-    )
-    val distressedIds = buildDistressedSet(households)
 
-    val mapped = households
-      .zip(financialStocks)
-      .map: (hh, stocks) =>
-        if hh.status == HhStatus.Bankrupt then
-          val bankruptHh = hh.copy(financialDistressState = HhFinancialDistressState.Bankruptcy)
-          (bankruptHh, None, stocks, MonthlyFlow.inactive(hh.id, stocks)) // absorbing barrier
-        else
-          val result = processHousehold(hh, stocks, world, rng, bankRates, equityIndexReturn, sectorWages, sectorVacancies, distressedIds, consumerCreditGate)
-          val flow   = MonthlyFlow(
-            householdId = hh.id,
-            openingDemandDeposit = stocks.demandDeposit,
-            openingConsumerLoan = stocks.consumerLoan,
-            monthlyIncome = result.income,
-            consumption = result.consumption,
-            rent = result.rent,
-            mortgageDebtService = result.debtService,
-            consumerApprovedOrigination = result.credit.newLoan,
-            consumerCreditDemand = result.credit.creditDemand,
-            consumerRejectedOrigination = result.credit.rejectedCreditDemand,
-            consumerBankRejectedOrigination = result.credit.bankRejectedCreditDemand,
-            consumerCreditCapacity = result.credit.creditCapacity,
-            consumerCreditAccessEligible = result.credit.creditAccessEligible,
-            liquidityShortfallFinancing = result.credit.liquidityShortfallFinancing,
-            consumerDebtService = result.credit.debtService,
-            consumerDefault = result.credit.defaultAmt,
-            consumerPrincipal = result.credit.principal,
-            closingConsumerLoan = result.financialStocks.consumerLoan,
-            consumerLoanDefault = result.credit.consumerLoanDefault,
-            liquidityBridgeChargeOff = result.credit.liquidityBridgeChargeOff,
-            unmetBasicConsumption = result.unmetBasicConsumption,
-            discretionaryConsumptionCompression = result.discretionaryConsumptionCompression,
-            consumptionShortfall = result.credit.liquidityShortfall.consumptionShortfall,
-            rentArrears = result.credit.liquidityShortfall.rentArrears,
-            mortgageArrears = result.credit.liquidityShortfall.mortgageArrears,
-            consumerDebtArrears = result.credit.liquidityShortfall.consumerDebtArrears,
-            temporaryOverdraft = result.credit.liquidityShortfall.temporaryOverdraft,
-          )
-          (result.newState, Some((hh.bankId, result)), result.financialStocks, flow)
-
-    val updated = mapped.map(_._1)
-    val stocks  = mapped.map(_._3)
-    val flows   = mapped.flatMap(_._2)
-    val monthly = mapped.map(_._4)
-    val totals  = { val t = StepTotals(); flows.foreach((_, r) => t.add(r)); t }
-    val agg     = computeAggregates(updated, stocks, marketWage, reservationWage, importAdj, totals)
-    val pbf     = if bankRates.isDefined then Some(buildPerBankFlows(flows, nBanks)) else None
-    require(
-      monthly.map(_.liquidityShortfallFinancing).sumPln == agg.totalLiquidityShortfallFinancing,
-      "Household.step monthly flow diagnostics must reconcile to aggregate liquidity shortfall financing",
-    )
-    require(
-      agg.totalLiquidityShortfallComponents == agg.totalLiquidityShortfallFinancing,
-      "Household.step shortfall components must reconcile to aggregate liquidity shortfall financing",
-    )
-    require(
-      agg.totalConsumerLoanDefault + agg.totalLiquidityBridgeChargeOff == agg.totalConsumerDefault,
-      "Household.step consumer default components must reconcile to aggregate consumer default",
-    )
-    StepResult(updated, agg, pbf, stocks, monthly)
-
-  /** Pre-compute distressed HH set for O(1) neighbor lookups. */
-  private def buildDistressedSet(households: Vector[State]): java.util.BitSet =
-    val bits = new java.util.BitSet(households.length)
-    var i    = 0
-    while i < households.length do
-      val hh = households(i)
-      hh.status match
-        case HhStatus.Bankrupt | HhStatus.Unemployed(_)                         => bits.set(i)
-        case _ if hh.financialDistressState != HhFinancialDistressState.Current =>
-          bits.set(i)
-        case _                                                                  =>
-      i += 1
-    bits
-
-  /** Sector mobility rate: fraction of employed in different sector than last.
-    */
-  private def sectorMobilityRate(updated: Vector[State]): Share =
-    val employed    = updated.flatMap: hh =>
-      hh.status match
-        case HhStatus.Employed(_, sec, _) => Some((hh.lastSectorIdx, sec))
-        case _                            => None
-    if employed.isEmpty then return Share.Zero
-    val crossSector = employed.count((last, cur) => last.toInt >= 0 && last != cur)
-    Share.fraction(crossSector, employed.length)
-
-  /** Base income, benefit, and updated status for one HH. */
-  private def computeIncome(hh: State)(using SimParams): (PLN, PLN, HhStatus) =
-    hh.status match
-      case HhStatus.Employed(firmId, sectorIdx, wage) =>
-        (wage, PLN.Zero, hh.status)
-      case HhStatus.Unemployed(months)                =>
-        val benefit = computeBenefit(months)
-        (benefit, benefit, HhStatus.Unemployed(months + 1))
-      case HhStatus.Retraining(monthsLeft, _, cost)   =>
-        (PLN.Zero, PLN.Zero, hh.status)
-      case HhStatus.Bankrupt                          =>
-        (PLN.Zero, PLN.Zero, HhStatus.Bankrupt)
-
-  /** Skill decay for long-term unemployed (onset after scarringOnset months).
-    */
-  private def applySkillDecay(hh: State, status: HhStatus)(using p: SimParams): Share =
-    status match
-      case HhStatus.Unemployed(months) if months >= p.household.scarringOnset =>
-        hh.skill * (Share.One - p.household.skillDecayRate)
-      case _                                                                  => hh.skill
-
-  /** Apply health scarring for long-term unemployed (cumulative, capped). */
-  private def applyHealthScarring(hh: State, status: HhStatus)(using p: SimParams): Share =
-    status match
-      case HhStatus.Unemployed(months) if months >= p.household.scarringOnset =>
-        (hh.healthPenalty + p.household.scarringRate).min(p.household.scarringCap)
-      case _                                                                  => hh.healthPenalty
-
-  /** Wage scar: accumulates during long-term unemployment, decays slowly once
-    * reemployed. Jacobson, LaLonde & Sullivan 1993; Davis & von Wachter 2011.
-    */
-  private def applyWageScar(hh: State, status: HhStatus)(using p: SimParams): Share =
-    status match
-      case HhStatus.Unemployed(months) if months >= p.household.scarringOnset =>
-        (hh.wageScar + p.household.wageScarRate).min(p.household.wageScarCap)
-      case _: HhStatus.Employed                                               =>
-        (hh.wageScar - p.household.wageScarDecay).max(Share.Zero)
-      case _                                                                  => hh.wageScar
-
-  /** State-dependent MPC: Carroll (1997) buffer-stock model.
-    *
-    * When savings/income > target → buffer is fat → MPC falls (more saving).
-    * When buffer depleted → MPC rises (spend everything). Unemployed get an
-    * additional boost (desperate spending from depleted buffers).
-    */
-  private[amorfati] def updateMpc(hh: State, financialStocks: FinancialStocks, income: PLN, status: HhStatus)(using p: SimParams): Share =
-    val baseMpc = hh.mpc
-    if income <= PLN.Zero then baseMpc
-    else
-      val targetSavings = income * p.household.bufferTargetMonths
-      val bufferRatio   = financialStocks.demandDeposit.ratioTo(targetSavings).toMultiplier
-      val deviation     = bufferRatio - Multiplier.One              // >0 = fat, <0 = depleted
-      val adjustment    = p.household.bufferSensitivity * deviation // Coefficient × Multiplier → Share
-      val bufferAdj     = (Share.One - adjustment).clamp(Share.Zero, Share.One)
-      val unemployedAdj = status match
-        case _: HhStatus.Unemployed => Share.One + p.household.mpcUnemployedBoost
-        case _                      => Share.One
-      (baseMpc * bufferAdj * unemployedAdj).clamp(MpcFloor, MpcCeiling)
-
-  /** Controlled drawdown from liquid savings buffers.
-    *
-    * Employed households only draw from savings above their target buffer.
-    * Stressed households may also draw from the lower half of the buffer, but
-    * still keep a protected floor.
-    */
-  private[amorfati] def computeSavingsDrawdown(
-      financialStocks: FinancialStocks,
-      cashOnHand: PLN,
-      status: HhStatus,
-  )(using p: SimParams): PLN =
-    val targetIncome  = cashOnHand.max(p.household.baseReservationWage)
-    val targetSavings = targetIncome * p.household.bufferTargetMonths
-    val protectedBuff = targetSavings * p.household.bufferProtectedShare
-    status match
-      case _: HhStatus.Unemployed | _: HhStatus.Retraining =>
-        (financialStocks.demandDeposit - protectedBuff).max(PLN.Zero) * p.household.bufferStressDrawdownRate
-      case _: HhStatus.Employed | HhStatus.Bankrupt        =>
-        (financialStocks.demandDeposit - targetSavings).max(PLN.Zero) * p.household.bufferExcessDrawdownRate
-
-  /** Fraction of social neighbors in distress (BitSet, O(k) per HH). */
-  private def neighborDistressRatioFast(hh: State, distressedIds: java.util.BitSet): Share =
-    if hh.socialNeighbors.isEmpty then Share.Zero
-    else
-      var count = 0
-      var i     = 0
-      while i < hh.socialNeighbors.length do
-        if distressedIds.get(hh.socialNeighbors(i).toInt) then count += 1
-        i += 1
-      Share.fraction(count, hh.socialNeighbors.length)
-
-  /** Public entry point for aggregate stats (used by BankingEconomics and
-    * tests). Flow totals default to zero — only distribution stats are
-    * computed.
-    */
+  /** Public entry point for household aggregate statistics. */
   def computeAggregates(
       households: Vector[State],
       financialStocks: Vector[FinancialStocks],
@@ -1565,225 +605,16 @@ object Household:
       retrainingAttempts: Int,
       retrainingSuccesses: Int,
   )(using SimParams): Aggregates =
-    computeAggregates(
+    HouseholdAggregateComputation.computeAggregates(
       households,
       financialStocks,
       marketWage,
       reservationWage,
       importAdj,
-      { val t = StepTotals(); t.retrainingAttempts = retrainingAttempts; t.retrainingSuccesses = retrainingSuccesses; t },
+      retrainingAttempts,
+      retrainingSuccesses,
     )
 
-  /** Aggregate stats: single-pass accumulation + sorted-array Gini/percentiles.
-    * Merges per-HH distribution stats with flow totals from StepTotals in one
-    * construction — no intermediate Aggregates + copy overwrite.
-    */
-  private def computeAggregates(
-      households: Vector[State],
-      financialStocks: Vector[FinancialStocks],
-      marketWage: PLN,
-      reservationWage: PLN,
-      importAdj: Share,
-      t: StepTotals,
-  )(using p: SimParams): Aggregates =
-    require(
-      households.length == financialStocks.length,
-      s"Household.computeAggregates requires aligned households and financialStocks, got ${households.length} households and ${financialStocks.length} stock rows",
-    )
-    val n = households.length
-
-    var nEmployed          = 0
-    var nUnemployed        = 0
-    var nRetraining        = 0
-    var nBankrupt          = 0
-    var nDistCurrent       = 0
-    var nDistStress        = 0
-    var nDistArrears       = 0
-    var nDistRestructuring = 0
-    var nDistDefaulted     = 0
-    var nDistBankruptcy    = 0
-    var sumSkill           = 0L
-    var sumHealth          = 0L
-    var sumSavings         = 0L
-    val incomes            = new Array[Long](n)
-    val consumptions       = new Array[Long](n)
-    val savingsArr         = new Array[Long](n)
-
-    // Hot path: O(N_hh) single-pass with mutable accumulators + in-place arrays.
-    // Intentionally imperative — foldLeft with 9-field accumulator would be slower and less readable.
-    var i = 0
-    while i < n do
-      val hh     = households(i)
-      val stocks = financialStocks(i)
-      hh.status match
-        case HhStatus.Employed(_, _, wage) =>
-          nEmployed += 1
-          incomes(i) = wage.toLong
-          sumSkill += hh.skill.toLong
-          sumHealth += hh.healthPenalty.toLong
-        case HhStatus.Unemployed(months)   =>
-          nUnemployed += 1
-          incomes(i) = computeBenefit(months).toLong
-          sumSkill += hh.skill.toLong
-          sumHealth += hh.healthPenalty.toLong
-        case HhStatus.Retraining(_, _, _)  =>
-          nRetraining += 1
-          incomes(i) = 0L
-          sumSkill += hh.skill.toLong
-          sumHealth += hh.healthPenalty.toLong
-        case HhStatus.Bankrupt             =>
-          nBankrupt += 1
-          incomes(i) = 0L
-
-      // Legacy Bankrupt status rows always count as Bankruptcy even if their
-      // separate financial-distress field predates this state machine.
-      val distressState =
-        if hh.status == HhStatus.Bankrupt then HhFinancialDistressState.Bankruptcy else hh.financialDistressState
-      distressState match
-        case HhFinancialDistressState.Current         => nDistCurrent += 1
-        case HhFinancialDistressState.LiquidityStress => nDistStress += 1
-        case HhFinancialDistressState.Arrears         => nDistArrears += 1
-        case HhFinancialDistressState.Restructuring   => nDistRestructuring += 1
-        case HhFinancialDistressState.Defaulted       => nDistDefaulted += 1
-        case HhFinancialDistressState.Bankruptcy      => nDistBankruptcy += 1
-
-      val rentRaw       = hh.monthlyRent.toLong
-      val mortgageRate  = p.monetary.initialRate + p.housing.mortgageSpread
-      val debtSvcRaw    = (scheduledMortgagePrincipal(stocks) + (stocks.mortgageLoan * mortgageRate.monthly)).toLong
-      val disposableRaw = math.max(0L, incomes(i) - rentRaw - debtSvcRaw)
-      consumptions(i) = FixedPointBase.multiplyRaw(disposableRaw, hh.mpc.toLong)
-      savingsArr(i) = stocks.demandDeposit.toLong
-      sumSavings += savingsArr(i)
-      i += 1
-
-    val nAlive = n - nBankrupt
-
-    // Sort each array once — reuse for Gini + percentiles + poverty
-    java.util.Arrays.sort(incomes)
-    java.util.Arrays.sort(savingsArr)
-    java.util.Arrays.sort(consumptions)
-
-    // Consumption split: flow totals (from StepTotals) are authoritative
-    val totalConsumption = t.goodsConsumption + t.rent
-    val importCons       = t.goodsConsumption * importAdj.min(ImportRatioCap)
-    val domesticCons     = totalConsumption - importCons
-
-    val medianIncomeRaw = if n > 0 then incomes(n / 2) else 0L
-
-    Aggregates(
-      employed = nEmployed,
-      unemployed = nUnemployed,
-      retraining = nRetraining,
-      bankrupt = nBankrupt,
-      totalIncome = t.income,
-      consumption = totalConsumption,
-      domesticConsumption = domesticCons,
-      importConsumption = importCons,
-      marketWage = marketWage,
-      reservationWage = reservationWage,
-      giniIndividual = giniSorted(incomes),
-      giniWealth = giniSorted(savingsArr),
-      meanSavings = if n > 0 then PLN.fromRaw(sumSavings / n) else PLN.Zero,
-      medianSavings = if n > 0 then PLN.fromRaw(savingsArr(n / 2)) else PLN.Zero,
-      povertyRate50 =
-        if n > 0 && medianIncomeRaw > 0L then Share.fraction(lowerBound(incomes, (PLN.fromRaw(medianIncomeRaw) * PovertyRate50Pct).toLong), n) else Share.Zero,
-      bankruptcyRate = if n > 0 then Share.fraction(nBankrupt, n) else Share.Zero,
-      meanSkill = if nAlive > 0 then Share.fromRaw(sumSkill / nAlive) else Share.Zero,
-      meanHealthPenalty = if nAlive > 0 then Share.fromRaw(sumHealth / nAlive) else Share.Zero,
-      retrainingAttempts = t.retrainingAttempts,
-      retrainingSuccesses = t.retrainingSuccesses,
-      consumptionP10 =
-        if n > 0 then PLN.fromRaw(consumptions((n * ConsumptionP10.toLong / com.boombustgroup.amorfati.fp.FixedPointBase.Scale).toInt)) else PLN.Zero,
-      consumptionP50 = if n > 0 then PLN.fromRaw(consumptions(n / 2)) else PLN.Zero,
-      consumptionP90 =
-        if n > 0 then PLN.fromRaw(consumptions(Math.min(n - 1, (n * ConsumptionP90.toLong / com.boombustgroup.amorfati.fp.FixedPointBase.Scale).toInt)))
-        else PLN.Zero,
-      meanMonthsToRuin = Scalar.Zero,
-      povertyRate30 =
-        if n > 0 && medianIncomeRaw > 0L then Share.fraction(lowerBound(incomes, (PLN.fromRaw(medianIncomeRaw) * PovertyRate30Pct).toLong), n) else Share.Zero,
-      totalRent = t.rent,
-      totalDebtService = t.debtService,
-      totalUnempBenefits = t.unempBenefits,
-      totalDepositInterest = t.depositInterest,
-      crossSectorHires = 0,
-      voluntaryQuits = t.voluntaryQuits,
-      sectorMobilityRate = sectorMobilityRate(households),
-      totalRemittances = t.remittances,
-      totalPit = t.pit,
-      totalSocialTransfers = t.socialTransfers,
-      totalConsumerDebtService = t.consumerDebtService,
-      totalConsumerOrigination = t.consumerOrigination,
-      totalConsumerApprovedOrigination = t.consumerApprovedOrigination,
-      totalLiquidityShortfallFinancing = t.liquidityShortfallFinancing,
-      totalConsumerDefault = t.consumerDefault,
-      totalConsumerPrincipal = t.consumerPrincipal,
-      totalConsumerCreditDemand = t.totalConsumerCreditDemand,
-      totalConsumerRejectedOrigination = t.totalConsumerRejectedOrigination,
-      totalConsumerBankRejectedOrigination = t.totalConsumerBankRejectedOrigination,
-      totalConsumerLoanDefault = t.consumerLoanDefault,
-      totalLiquidityBridgeChargeOff = t.liquidityBridgeChargeOff,
-      totalUnmetBasicConsumption = t.unmetBasicConsumption,
-      totalDiscretionaryConsumptionCompression = t.discretionaryConsumptionCut,
-      totalConsumptionShortfall = t.consumptionShortfall,
-      totalRentArrears = t.rentArrears,
-      totalMortgageArrears = t.mortgageArrears,
-      totalConsumerDebtArrears = t.consumerDebtArrears,
-      totalTemporaryOverdraft = t.temporaryOverdraft,
-      totalMortgagePrincipal = t.mortgagePrincipal,
-      totalMortgageInterest = t.mortgageInterest,
-      distressCurrent = nDistCurrent,
-      distressLiquidityStress = nDistStress,
-      distressArrears = nDistArrears,
-      distressRestructuring = nDistRestructuring,
-      distressDefaulted = nDistDefaulted,
-      distressBankruptcy = nDistBankruptcy,
-    )
-
-  /** Gini coefficient for a pre-sorted array (handles negatives by shifting).
-    */
+  /** Gini coefficient for a pre-sorted array. */
   def giniSorted(sorted: Array[Long]): Share =
-    try giniSortedLong(sorted)
-    catch case _: ArithmeticException => giniSortedBig(sorted)
-
-  private def giniSortedLong(sorted: Array[Long]): Share =
-    val n           = sorted.length
-    if n <= 1 then return Share.Zero
-    val minVal      = sorted(0)
-    if minVal == Long.MinValue then throw new ArithmeticException("gini shift overflows Long")
-    val shift       = if minVal < 0L then -minVal else 0L
-    var total       = 0L
-    var weightedSum = 0L
-    var i           = 0
-    while i < n do
-      val v      = java.lang.Math.addExact(sorted(i), shift)
-      total = java.lang.Math.addExact(total, v)
-      val weight = 2L * (i.toLong + 1L) - n.toLong - 1L
-      weightedSum = java.lang.Math.addExact(weightedSum, java.lang.Math.multiplyExact(weight, v))
-      i += 1
-    if total <= 0L then Share.Zero
-    else Share.fromRaw(FixedPointBase.ratioRaw(weightedSum, java.lang.Math.multiplyExact(n.toLong, total)))
-
-  private def giniSortedBig(sorted: Array[Long]): Share =
-    val n           = sorted.length
-    if n <= 1 then return Share.Zero
-    val minVal      = sorted(0)
-    val shift       = if minVal < 0L then BigInt(minVal).abs else BigInt(0)
-    var total       = BigInt(0)
-    var weightedSum = BigInt(0)
-    var i           = 0
-    while i < n do
-      val v = BigInt(sorted(i)) + shift
-      total += v
-      weightedSum += BigInt(2L * (i.toLong + 1L) - n.toLong - 1L) * v
-      i += 1
-    if total <= 0 then Share.Zero else Share.fromRaw(scaledDivRaw(weightedSum, BigInt(n) * total))
-
-  /** Binary search: count of elements < threshold in a sorted array. */
-  private def lowerBound(sorted: Array[Long], threshold: Long): Int =
-    var lo = 0
-    var hi = sorted.length
-    while lo < hi do
-      val mid = (lo + hi) >>> 1
-      if sorted(mid) < threshold then lo = mid + 1
-      else hi = mid
-    lo
+    HouseholdAggregateComputation.giniSorted(sorted)
